Use PrismAsyncLight from react-syntax-highlighter to save 139K gzipped
This component defers loading most of the syntax highlighting code until
after the page loads, and it will only download syntaxes that are used on
the page.

It comes with a tradeoff though: syntax highlighting no longer works with
client JS disabled, i.e. the code blocks sent from the server are not
highlighted.